deduplicate migrate and v2upgrade code
